Cleanup graph names for graphs with two metrics.
When graphing two metrics, the name can get long and messy (i.e.
"metric1 and metric2"). Instead, only display the general topic.
e.g. If metric1 and metric2 are both CPU related metrics, find
the common topic of CPU and use that as the name of the graph.
Relates to OS-207.

Change-Id: Id6b7d45dc54f10878182b1859e3d83b19fb4c6be